phase 2
  - Before: 1949 L in one file.
  - After: 14 category files (each co-locating sigs + docs for its domain),
  one macros file, one tests file, one 86-L aggregator.
  - builtins.rs went 1949 → 86 L (23× reduction). Adding a new builtin is now
   a two-line edit in a single file.
  - Public API: identical (builtin_signature, builtin_signatures,
  builtin_doc, builtin_docs).
  - Duplicated category banners in the old docs
  (TCP/OS/Regex/Encoding/Crypto/Compression/Signal/Terminal) are now unified
  within their respective category files — the two formerly-separate blocks
  sit next to each other and could be merged in a follow-up pass.
  - 341 lib tests pass.
